budget mission et pagination request

+++ BACKEND/rh-backend/src/MyApp.Api/Repositories/mission/MissionBudgetRepository.cs
@@ -1,4 +1,5 @@
-﻿using MyApp.Api.Data;
+﻿using Microsoft.EntityFrameworkCore;
+using MyApp.Api.Data;
 using MyApp.Api.Entities.mission;
 
 namespace MyApp.Api.Repositories.mission
@@ -10,7 +11,7 @@ namespace MyApp.Api.Repositories.mission
         Task<MissionBudget?> GetByDirectionNameAsync(string directionName);
         Task AddAsync(MissionBudget entity);
         Task UpdateAsync(MissionBudget entity);
-        Task DeleteAsync(MissionBudget entity);
+        Task DeleteAsync(string id);
     }
 
     public class MissionBudgetRepository : IMissionBudgetRepository
@@ -24,20 +25,27 @@ namespace MyApp.Api.Repositories.mission
 
         public async Task<IEnumerable<MissionBudget>> GetAllAsync()
         {
-            return await Task.FromResult(_context.MissionBudgets.ToList());
+            return await _context.MissionBudgets
+                .Include(m => m.User)
+                .OrderByDescending(m => m.CreatedAt)
+                .ToListAsync();
         }
 
         public async Task<MissionBudget?> GetByIdAsync(string id)
         {
-            return await _context.MissionBudgets.FindAsync(id);
+            return await _context.MissionBudgets
+                        .AsNoTracking()
+                        .FirstOrDefaultAsync(m => m.MissionBudgetId == id);
         }
 
-        public async Task<MissionBudget?> GetByDirectionNameAsync(string directionName) // 👈 ajout
+        public async Task<MissionBudget?> GetByDirectionNameAsync(string directionName)
         {
-            return await Task.FromResult(
-                _context.MissionBudgets.FirstOrDefault(m => m.DirectionName == directionName)
-            );
+            return await _context.MissionBudgets
+                .Where(m => m.DirectionName == directionName)
+                .OrderByDescending(m => m.CreatedAt)
+                .FirstOrDefaultAsync();
         }
+
 
         public async Task AddAsync(MissionBudget entity)
         {
@@ -51,9 +59,10 @@ namespace MyApp.Api.Repositories.mission
             await _context.SaveChangesAsync();
         }
 
-        public async Task DeleteAsync(MissionBudget entity)
+        public async Task DeleteAsync(string id)
         {
-            _context.MissionBudgets.Remove(entity);
+            var entity = await _context.MissionBudgets.FindAsync(id);
+            if (entity != null) _context.MissionBudgets.Remove(entity);
             await _context.SaveChangesAsync();
         }
     }
